requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
mysql-connector-python==8.0.26
flask-cors==3.0.10